+++ spec/models/place_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe Place, type: :model do
   end
 
   describe 'enums' do
-    it { is_expected.to define_enum_for(:source).with_values(%i[manual photon]) }
+    it { is_expected.to define_enum_for(:source).with_values(%i[manual photon gpx_waypoint geojson_point]) }
   end
 
   describe 'scopes' do